End of ex. 3

+++ src/discovering/PWM.java
@@ -54,7 +54,7 @@ public class PWM {
 		return s+"\n";
 	}
 	
-	public int getScore(){
+	public double getScore(){
 		return this.consensusScore();
 	}
